+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 ## Changelog
 
+### 10.5.0
+
+* `[fmtc]` Added new methods `LPrintf`, `LPrintln`, `TLPrintf` and `TLPrintln`
+* `[fmtc]` Fixed bug with parsing reset and modification tags (_found by go-fuzz_)
+* `[fmtc]` Code refactoring
+
 ### 10.4.0
 
 * `[fmtc]` Improved work with temporary output (`TPrintf`, `TPrintln`)